chore(benchmarks): remove duplicate rank-one task
