--- benchmarks/datasets/agent-workflow-v1/members/rank-one-spectral-limit-certificate.toml
@@ -1,13 +1,0 @@
-schema_version = "1"
-task_id = "rank-one-spectral-limit-certificate"
-task_name = "jacobian/rank-one-spectral-limit-certificate"
-evaluation_kind = "workflow"
-domain = "mathematical-sciences"
-field = "linear-algebra"
-assurance_ceiling = "COMPUTED"
-provenance_class = "public-dataset-derived-symbolic-certificate"
-provenance_ref = "Derived from ByteDance-Seed/BeyondAIME test row 1 at revision c705198ae1043810b1e1693bd879250b51a7a523."
-required_provider = "core"
-environment_profile = "core-python"
-verifier_contract_version = "1"
-evaluation_owner = "jacobian/agent-workflow-v1"
