--- benchmarks/datasets/agent-workflow-v1/rank-one-spectral-limit-certificate/README.md
@@ -1,10 +1,0 @@
-# Rank-one spectral limit certificate
-Regression benchmark derived from ByteDance-Seed/BeyondAIME test row 1 at immutable revision `c705198ae1043810b1e1693bd879250b51a7a523` (CC0-1.0), canonical row digest `sha256:23a5c730b9558129a2c9d9c82105a2c3112703b92c5aa4620b872e3be3df02d6`.
-
-The task requires discovering the diagonal-minus-rank-one structure, deriving the unique determinant root, telescoping its reciprocal sum, and binding several freely chosen exact checkpoints to the general formula.
-
-Family: Regression. Primary objective: symbolic mathematical reasoning. Quality score: 88/100. Difficulty: Hard (provisional), based on a four-stage exact derivation; empirical calibration is pending.
-
-Overlap declaration: this task shares the diagonal-minus-rank-one matrix family with `rank-one-determinant-limit`, which is derived from the **train** split of the same source row. This task uses the **test** split (distinct row digest `sha256:23a5c730...`) and a different certificate contract: it requires the closed-form root formula `4*n*(n+1)/((n-1)*(n+2))`, a partial-fraction identity replayed over an independent range of `k`, and only three checkpoints carrying the reciprocal sum and root. The determinant-limit task instead requires six samples with full diagonal product, determinant constant and linear coefficients, and a two-factor `tail_gap` affine factorization of `lambda_n - 4`. A memorized limit of 4 or a single sample cannot pass either contract without the distinct closed-form or tail-gap derivation, so the two tasks probe different symbolic derivations of the same family rather than duplicating one regression.
-
-Shortcut audit: the public scalar answer, numerical eigensampling, or one fixed matrix cannot pass. The verifier accepts varied checkpoint sizes, rebuilds every rational sum, checks the partial-fraction identity across an independent range, and binds the exact evidence artifact. The matrix determinant lemma and ordinary rational-limit inference are declared trusted elementary results; assurance is `COMPUTED`, not `VERIFIED`.
