--- benchmarks/datasets/agent-workflow-v1/rank-one-spectral-limit-certificate/environment/input.json
@@ -1,1 +1,0 @@
-{"schema_version":"1","task_id":"jacobian/rank-one-spectral-limit-certificate","source":{"dataset":"ByteDance-Seed/BeyondAIME","revision":"c705198ae1043810b1e1693bd879250b51a7a523","split":"test","row":1,"license":"CC0-1.0","row_digest":"sha256:23a5c730b9558129a2c9d9c82105a2c3112703b92c5aa4620b872e3be3df02d6"},"matrix":{"indices":"k=2,...,n","diagonal_base":"k^3-k","all_entries_subtract":"lambda"},"checkpoint_bounds":[4,30],"minimum_checkpoints":3}
