--- benchmarks/datasets/agent-workflow-v1/rank-one-spectral-limit-certificate/environment/submission_schema.json
@@ -1,1 +1,0 @@
-{"$schema":"https://json-schema.org/draft/2020-12/schema","type":"object","additionalProperties":false,"required":["task_id","conclusion","result","claimed_assurance","scope","completeness","evidence","limitations"],"properties":{"task_id":{"const":"jacobian/rank-one-spectral-limit-certificate"},"conclusion":{"const":"LIMIT_DERIVED_EXACTLY"},"result":{"$ref":"#/$defs/result"},"claimed_assurance":{"enum":["UNVERIFIED","COMPUTED"]},"scope":{"const":"DECLARED_MATRIX_FAMILY"},"completeness":{"const":"COMPLETE"},"evidence":{"type":"array","minItems":1,"maxItems":1,"items":{"type":"object","additionalProperties":false,"required":["path","sha256"],"properties":{"path":{"const":"evidence/spectral-certificate.json"},"sha256":{"type":"string","pattern":"^sha256:[0-9a-f]{64}$"}}}},"limitations":{"const":["MATRIX_DETERMINANT_LEMMA_TRUSTED","RATIONAL_LIMIT_INFERENCE_NOT_PROOF_ASSISTANT_VERIFIED"]}},"$defs":{"rat":{"type":"object","additionalProperties":false,"required":["numerator","denominator"],"properties":{"numerator":{"type":"integer"},"denominator":{"type":"integer","minimum":1}}},"checkpoint":{"type":"object","additionalProperties":false,"required":["n","reciprocal_sum","root"],"properties":{"n":{"type":"integer","minimum":4,"maximum":30},"reciprocal_sum":{"$ref":"#/$defs/rat"},"root":{"$ref":"#/$defs/rat"}}},"result":{"type":"object","additionalProperties":false,"required":["rank_one_sign","partial_fraction","checkpoints","root_formula","limit"],"properties":{"rank_one_sign":{"const":"DIAGONAL_MINUS_LAMBDA_ONES"},"partial_fraction":{"type":"object","additionalProperties":false,"required":["scale","left_coefficient","right_coefficient"],"properties":{"scale":{"$ref":"#/$defs/rat"},"left_coefficient":{"type":"integer","minimum":-100,"maximum":100},"right_coefficient":{"type":"integer","minimum":-100,"maximum":100}}},"checkpoints":{"type":"array","minItems":3,"maxItems":8,"items":{"$ref":"#/$defs/checkpoint"}},"root_formula":{"type":"string","minLength":1,"description":"Exact rational expression for the determinant root as a function of n."},"limit":{"$ref":"#/$defs/rat"}}}}}
