--- benchmarks/datasets/agent-workflow-v1/rank-one-spectral-limit-certificate/solution/solve.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-set -eu
-mkdir -p /app/evidence
-cp /solution/spectral-certificate.json /app/evidence/spectral-certificate.json
-cp /solution/submission.json /app/submission.json
