--- benchmarks/datasets/agent-workflow-v1/rank-one-spectral-limit-certificate/solution/spectral-certificate.json
@@ -1,1 +1,0 @@
-{"schema_version":"1","task_id":"jacobian/rank-one-spectral-limit-certificate","result":{"rank_one_sign":"DIAGONAL_MINUS_LAMBDA_ONES","partial_fraction":{"scale":{"numerator":1,"denominator":2},"left_coefficient":1,"right_coefficient":-1},"checkpoints":[{"n":4,"reciprocal_sum":{"numerator":9,"denominator":40},"root":{"numerator":40,"denominator":9}},{"n":7,"reciprocal_sum":{"numerator":27,"denominator":112},"root":{"numerator":112,"denominator":27}},{"n":12,"reciprocal_sum":{"numerator":77,"denominator":312},"root":{"numerator":312,"denominator":77}}],"root_formula":"4*n*(n+1)/((n-1)*(n+2))","limit":{"numerator":4,"denominator":1}},"limitations":["MATRIX_DETERMINANT_LEMMA_TRUSTED","RATIONAL_LIMIT_INFERENCE_NOT_PROOF_ASSISTANT_VERIFIED"]}
