--- benchmarks/datasets/agent-workflow-v1/rank-one-spectral-limit-certificate/tests/expected.json
@@ -1,1 +1,0 @@
-{"task_id":"jacobian/rank-one-spectral-limit-certificate","conclusion":"LIMIT_DERIVED_EXACTLY","maximum_assurance":"COMPUTED"}
