--- benchmarks/datasets/agent-workflow-v1/rank-one-spectral-limit-certificate/tests/test.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-set -eu
-python /tests/verifier.py
